fix(Book): Arreglar bloque de días por registro de reserva

+++ Assert.Application/Interfaces/IAppBookService.cs
@@ -8,7 +8,7 @@ namespace Assert.Application.Interfaces
         Task<ReturnModelDTO<PayPriceCalculationDTO>> CalculatePrice(long listingRentId, DateTime startDate, DateTime endDate, int guestId,
            Dictionary<string, string> clientData, bool useTechnicalMessages);
 
-        Task<ReturnModelDTO<int>> UpsertBookAsync(BookDTO incomingBook);
+        Task<ReturnModelDTO<long>> UpsertBookAsync(BookDTO incomingBook);
 
         Task<ReturnModelDTO<BookDTO>> GetBookByIdAsync(long bookId);
